improve strorage calcs, tg messages

+++ main.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
-	unix "golang.org/x/sys/unix"
+	"syscall"
 	
 	"github.com/tidwall/gjson"
 	"github.com/yanzay/tbot/v2"
@@ -16,48 +16,60 @@ var (
 	client *tbot.Client
 )
 
+func calcUsedDiskVolume(path string) uint64 {
+	fs := syscall.Statfs_t{}
+	err := syscall.Statfs(path, &fs)
+	if err != nil {
+		panic(err)
+	}
+
+	all := fs.Blocks * uint64(fs.Bsize)
+	free := fs.Bfree * uint64(fs.Bsize)
+	used := all - free
+	return used*100/all
+}
+
 func main() {
-	var (
-		configPath string
-		stat unix.Statfs_t
-	)
+	var configPath string
 
 	flag.StringVar(&configPath, "config", "./config.json", "Path to a config file")
 	flag.Parse()
 
 	content, err := ioutil.ReadFile(configPath) 
 	if err != nil {
-    fmt.Println(err)
-  }
-	
-	token := gjson.Get(string(content), "telegram-bot.token")
-	bot = tbot.New(token.Str)
-	client = bot.Client()
-
-	wd, err := os.Getwd()
-	unix.Statfs(wd, &stat)
-
-	// Available blocks * size per block = available space in bytes
-	fmt.Println(stat.Bavail * uint64(stat.Bsize))
-	
-	folders := gjson.Get(string(content), "folders")
-	for _, folder := range folders.Array() {
-		err = os.RemoveAll(folder.String())
-		if err != nil {
-			fmt.Println(err)
-		}
-		fmt.Println(folder, "was removed")
-
-		err = os.Mkdir(folder.String(), 0755)
-		if err != nil {
-			fmt.Println(err)
-		}
-		fmt.Println(folder, "was created again")
+		fmt.Println(err)
 	}
 	
-	bot.HandleMessage("/start", func(m *tbot.Message) {
-		client.SendMessage(m.Chat.ID, "hi")
-	})
+	token := gjson.Get(string(content), "telegram-bot.token").Str
+	channel := gjson.Get(string(content), "telegram-bot.channel").Str
+	folders := gjson.Get(string(content), "folders").Array()
+	maxVolume := gjson.Get(string(content), "maxVolume").Uint()
+	bot = tbot.New(token)
+	client = bot.Client()
+	
+	usedDiskBefore := calcUsedDiskVolume("/")
+	
+	if maxVolume < usedDiskBefore {
+		for _, folder := range folders {
+			err = os.RemoveAll(folder.String())
+			if err != nil {
+				fmt.Println(err)
+			}
+			fmt.Println(folder, "was removed")
+
+			err = os.Mkdir(folder.String(), 0755)
+			if err != nil {
+				fmt.Println(err)
+			}
+			fmt.Println(folder, "was created again")
+		}
+		usedDiskAfter := calcUsedDiskVolume("/")
+
+		client.SendMessage(channel, fmt.Sprintf("Объём занимаемого места: %d%%.\nПосле очистки: %d%%", usedDiskBefore, usedDiskAfter))
+	} else {
+		// fmt.Println("!=")
+		client.SendMessage(channel, fmt.Sprintf("Объём занимаемого места: %d%%.\nОчистка не проводилась", usedDiskBefore))
+	}
 	err = bot.Start()
 	if err != nil {
 		panic(err)
